fix(raffle): bound the winner pick; land the pure-mechanics split
pickWinners clamps n to the same ceiling Open enforces before any
narrowing conversion — the count arrives from chat args or stored JSON,
and CodeQL rightly flags an unchecked int64->int. Also moves the pure
helpers (clamp, key building, pick, digest, text shaping) into
raffle_mechanics.go for real: the earlier split commit lost the bodies
to a bad cut and shipped only comment fragments.
